Crate pallet_subscriptions

Subscription Module

The Subscription module allows supporters of any space that represents a creator or a nonprofit organization to contribute financially (with native tokens) a daily, weekly, monthly, quarterly or yearly basis. This module allows content creators to monetize their content via recurring payments from their supporters.

This pallet provides a way for creators to create a list of subscription plans (aka levels, tiers) and specify a different price and time period per each plan. There are several pre-built subscription periods: Daily, Weekly, Monthly, Quarterly and Yearly.

This pallet uses Substrate’s Schedule pallet to schedule recurring transfers from supporters’ (patrons’) wallets to creators’ wallets.

Modules

functions

Structs

Module
The module declaration.

NextPlanId
The next subscription plan id.

NextSubscriptionId

PlanById

PlanIdsBySpace

RecipientWallet
A recipient’s wallet that receives transfers sent from their subscribers.

SubscriberWallet
A subscriber’s wallet that is used to pay for their active subscriptions.

Subscription

SubscriptionById

SubscriptionIdsByPatron

SubscriptionIdsBySpace

SubscriptionPlan

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

SubscriptionPeriod

Constants

FIRST_SUBSCRIPTION_PLAN_ID

Traits

Trait
The pallet’s configuration trait.

Type Definitions

Event
RawEvent specialized for the configuration Trait

SubscriptionId

SubscriptionPlanId